# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ariadne Daemon API",
    "description": "REST API of ariadned, the coding-agent orchestrator daemon.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/v1/repositories": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "List repositories.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Create a repository.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDto"
                }
              }
            }
          },
          "400": {
            "description": "not an absolute path, not a git work tree, an unknown branch, or a forge pin that is no catalog model"
          },
          "409": {
            "description": "this path and base branch are already registered, `ai` was asked for while the AI permission model is off, or the forge integration cannot be enabled"
          }
        }
      }
    },
    "/v1/repositories/{id}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Get a repository.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "repositories"
        ],
        "summary": "Update a repository.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDto"
                }
              }
            }
          },
          "400": {
            "description": "not an absolute path, not a git work tree, an unknown branch, or a forge pin that is no catalog model"
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "this path and base branch are already registered, `ai` was asked for while the AI permission model is off, or the forge integration cannot be enabled"
          }
        }
      },
      "delete": {
        "tags": [
          "repositories"
        ],
        "summary": "Delete a repository.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRepositoryRequest": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "base_branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "Omit for the repo's currently checked-out branch."
          },
          "default_landing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Landing",
                "description": "Omit for `merge`."
              }
            ]
          },
          "default_workflow": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "forge": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ForgeUpdate",
                "description": "The forge integration to set up once the remote is detected."
              }
            ]
          },
          "path": {
            "type": "string",
            "description": "Absolute path of an existing git work tree.",
            "example": "/home/me/dev/ariadne"
          },
          "permission_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PermissionMode",
                "description": "Omit for `auto`."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ForgeDto": {
        "type": "object",
        "description": "The forge a repository's remote is on, and whether Ariadne works with it.",
        "required": [
          "webhook",
          "kind",
          "host",
          "owner",
          "name",
          "remote",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "description": "Lower-cased, like `owner` and `name`.",
            "example": "github.com"
          },
          "kind": {
            "$ref": "#/components/schemas/ForgeKind"
          },
          "login": {
            "type": [
              "string",
              "null"
            ],
            "description": "The account the forge CLI is signed in as, stored on enable."
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "remote": {
            "type": "string",
            "description": "The remote it was read off, `origin` where there is one."
          },
          "review_effort": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_model": {
            "type": [
              "string",
              "null"
            ],
            "description": "The pin of the session that reviews a request; null starts none."
          },
          "webhook": {
            "$ref": "#/components/schemas/WebhookDto"
          }
        }
      },
      "ForgeKind": {
        "type": "string",
        "description": "The forge a repository's remote is on (025): which CLI speaks to it.",
        "enum": [
          "github",
          "gitlab"
        ]
      },
...
      "ForgeUpdate": {
        "type": "object",
        "description": "A change to the forge integration; absent fields stay unchanged. A model\nwritten empty clears that role's pin and its effort.",
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "review_effort": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_model": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "Landing": {
        "type": "string",
        "description": "How the tasks of one goal end.\n\nThe one thing about the end of a task the author has to be told, since the\ncommands it runs differ entirely between the endings. The user chooses it\nonce, for the whole goal, when the goal is created, and every task of the\ngoal follows it: some work lands on the base branch, some goes through a\nrequest the author then sees to its merge, and some has nothing to land at\nall — a report filed, a document published, a release cut. Every ending\nreaches [`TaskStatus::Finished`]; landing is one way of getting there\nrather than the meaning of being there.\n\nWhich forge a published request goes to is *not* here: `origin` says\nwhether it is GitHub or GitLab, and asking the remote at landing time\ncannot go stale the way a second copy of the answer would.",
        "enum": [
          "merge",
          "pull_request",
          "none",
          "feature_branch"
        ]
      },
...
      "PermissionMode": {
        "type": "string",
        "description": "How the ACP runtime answers a tool permission request.",
        "enum": [
          "auto",
          "ask",
          "learn",
          "ai"
        ]
      },
...
      "RepositoryDto": {
        "type": "object",
        "required": [
          "id",
          "path",
          "base_branch",
          "permission_mode",
          "default_landing",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "base_branch": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "default_landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "The landing a new goal uses where its request leaves landing out."
          },
          "default_workflow": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "forge": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ForgeDto",
                "description": "The forge its remote is on, or null where the checkout has no usable\nremote (025)."
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Absolute path of the checkout."
          },
          "permission_mode": {
            "$ref": "#/components/schemas/PermissionMode",
            "description": "How the ACP permission requests of every session in this checkout are\nanswered."
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "UpdateRepositoryRequest": {
        "type": "object",
        "description": "Partial update; absent fields stay unchanged.",
        "properties": {
          "base_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_landing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Landing",
                "description": "Absent = unchanged."
              }
            ]
          },
          "default_workflow": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "New description, or empty to clear it. Absent = unchanged."
          },
          "forge": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ForgeUpdate",
                "description": "Absent = unchanged."
              }
            ]
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "permission_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PermissionMode",
                "description": "Absent = unchanged."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "WebhookDto": {
        "type": "object",
        "description": "Public hook status. The hook secret is never serialized.",
        "required": [
          "state"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "fetch_error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the last fetch of the repository's requests failed, or null once\none works: what says polling works, where no hook is live."
          },
          "last_delivery_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}